pyproject.toml
[tool.poetry]
name = "educe"
version = "2.0.0"
description = "Create colorschemes from images"
authors = ["Purple Myst <PurpleMyst@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.6"
pystache = "^0.5.4"
PyYAML = "^5.1"
Pillow = "^5.4"
termcolor = "^1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
